restructuring the template
