+++ src/app/app.module.ts
@@ -8,7 +8,7 @@ import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { HttpClient } from '@angular/common/http';
 import { AppComponent } from './app.component';
-import { SampleDataModule } from './sampledata/sampledata.module';
+import { SampleDataModule } from './home/sampledata/sampledata.module';
 import { EffectsModule } from '@ngrx/effects';
 import { StoreModule, MetaReducer } from '@ngrx/store';
 import { reducers, CustomSerializer } from './store';
