+++ src/app/auth/components/login.component.html
@@ -1,6 +1,6 @@
 <div class="login-container">
   <mat-toolbar class="public-header-container" color="primary">
-    <public-header [sideNavOpened]="true"></public-header>
+    <public-nav-bar [sideNavOpened]="true"></public-nav-bar>
   </mat-toolbar>
 <div>
     <mat-card class="login-card">
