+++ src/app/home/sampledata/components/sampledata-alert/sampledata-alert.component.scss
