+++ src/app/home/sampledata/components/sampledata-grid/sampledata-grid.component.scss
