+++ src/app/home/sampledata/sampledata.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CoreModule } from '../core/core.module';
+import { CoreModule } from '../../core/core.module';
 import { TranslateModule } from '@ngx-translate/core';
 import { StoreModule } from '@ngrx/store';
 import { reducers } from './store/reducers/index';
 import { EffectsModule } from '@ngrx/effects';
 import { SampleDataDialogComponent } from '../sampledata/components/sampledata-dialog/sampledata-dialog.component';
 import { SampleDataRoutingModule } from './sampledata-routing.module';
-import { HomeModule } from '../home/home.module';
+import { HomeModule } from '../home.module';
 
 import { SampleDataGridComponent } from './components/sampledata-grid/sampledata-grid.component';
 import { SampleDataAlertComponent } from './components/sampledata-alert/sampledata-alert.component';
