+++ src/app/home/sampledata/store/effects/index.ts
@@ -1,0 +1,3 @@
+import { SampleDataEffects } from './sampledata.effects';
+export const sampleDataEffects: any[] = [SampleDataEffects];
+export * from './sampledata.effects';
