+++ src/app/home/sampledata/store/reducers/index.ts
