+++ src/app/layout/layout.module.ts
@@ -6,6 +6,7 @@ import { CommonModule } from '@angular/common';
 import { AppRoutingModule } from '../app-routing.module';
 
 import { HeaderComponent } from './header/header.component';
+import { NavBarComponent} from './nav-bar/nav-bar.component';
 
 /* @export
  * @class LayoutModule
@@ -13,7 +14,7 @@ import { HeaderComponent } from './header/header.component';
 @NgModule({
   imports: [CommonModule, CoreModule, AppRoutingModule, TranslateModule],
   providers: [],
-  declarations: [HeaderComponent],
-  exports: [HeaderComponent],
+  declarations: [NavBarComponent, HeaderComponent],
+  exports: [NavBarComponent],
 })
 export class LayoutModule {}
