+++ src/app/layout/nav-bar/nav-bar.component.scss
