--- src/app/sampledata/store/effects/index.ts
@@ -1,3 +1,0 @@
-// import { SampleDataEffects } from './sampledata.effects';
-// export const sampleDataEffects: any[] = [SampleDataEffects];
-// export * from './sampledata.effects';
